+++ backend/de.metas.cucumber/src/test/java/de/metas/cucumber/stepdefs/process/AD_Process_Create_StepDef.java
@@ -1,4 +1,4 @@
-package de.metas.cucumber.stepdefs;
+package de.metas.cucumber.stepdefs.process;
 
 import de.metas.cucumber.stepdefs.DataTableRow;
 import de.metas.cucumber.stepdefs.DataTableRows;
